maint #343 extract solver discovery from misc.py into solver_utils.py
Create hklpy2/solver_utils.py with SOLVER_ENTRYPOINT_GROUP, get_solver,
solver_factory, and solvers.  Remove all four from misc.py (no re-export
shim).  Update __init__.py, ops.py, diffract.py (deferred import inside
creator()), and all test files to import from solver_utils directly.

Agent: OpenCode (claudesonnet46)
